+++ classes/task/run_reports.php
@@ -61,6 +61,7 @@ class run_reports extends \core\task\scheduled_task {
 
         list($startofthisweek, $startoflastweek) = report_customsql_get_week_starts($timenow);
         list($startofthismonth) = report_customsql_get_month_starts($timenow);
+        $startofthishour = report_customsql_get_hour_starts($timenow);
 
         mtrace("... Looking for old temp CSV files to delete.");
         $numdeleted = report_customsql_delete_old_temp_files($startoflastweek);
@@ -71,12 +72,18 @@ class run_reports extends \core\task\scheduled_task {
         // Get daily scheduled reports.
         $dailyreportstorun = report_customsql_get_ready_to_run_daily_reports($timenow);
 
-        // Get weekly and monthly scheduled reports.
-        $scheduledreportstorun = $DB->get_records_select('report_customsql_queries',
-                                            "(runable = 'weekly' AND lastrun < :startofthisweek) OR
-                                             (runable = 'monthly' AND lastrun < :startofthismonth)",
-                                            array('startofthisweek' => $startofthisweek,
-                                                  'startofthismonth' => $startofthismonth), 'lastrun');
+        // Get other scheduled reports.
+        // 2019.08.27.00 - Add hourly and always reports.
+        $scheduledreportstorun = $DB->get_records_select(
+            'report_customsql_queries',
+            "(runable = 'weekly' AND lastrun < :startofthisweek) OR
+             (runable = 'monthly' AND lastrun < :startofthismonth) OR
+             (runable = 'hourly' AND lastrun < :startofthishour) OR
+             runable = 'always'",
+            array('startofthisweek' => $startofthisweek,
+                'startofthismonth' => $startofthismonth,
+                'startofthishour' => $startofthishour),
+            'lastrun');
 
         // All reports ready to run.
         $reportstorun = array_merge($dailyreportstorun, $scheduledreportstorun);