+++ locallib.php
@@ -288,6 +288,8 @@ function report_customsql_runable_options($type = null) {
         return array('manual' => get_string('manual', 'report_customsql'));
     }
     return array('manual' => get_string('manual', 'report_customsql'),
+                 'always' => get_string('automaticallyalways', 'report_customsql'),
+                 'hourly' => get_string('automaticallyhourly', 'report_customsql'),
                  'daily' => get_string('automaticallydaily', 'report_customsql'),
                  'weekly' => get_string('automaticallyweekly', 'report_customsql'),
                  'monthly' => get_string('automaticallymonthly', 'report_customsql')
@@ -556,8 +558,21 @@ function report_customsql_get_month_starts($timenow) {
     );
 }
 
+// 2019.08.27.00
+function report_customsql_get_hour_starts($timenow) {
+    $dateparts = getdate($timenow);
+    return mktime($dateparts['hours'], 0, 0, $dateparts['mon'],
+        $dateparts['mday'], $dateparts['year']);
+}
+
+// 2019.08.27.00
+// Add always and hourly cases.
 function report_customsql_get_starts($report, $timenow) {
     switch ($report->runable) {
+        case 'always':
+            return $timenow;
+        case 'hourly':
+            return report_customsql_get_hour_starts($timenow);
         case 'daily':
             return report_customsql_get_daily_time_starts($timenow, $report->at);
         case 'weekly':